account for uncovered URL formats
Sometimes an image URL can come as `//domain.../image.jpg`(https://stackoverflow.com/questions/4831741/can-i-change-all-my-http-links-to-just) which would fail if we try to access 'scheme'

Sometimes the URL can come with a port different from `:80` or `:443`, which would fail if we only account for `host` and `path`.

This change would remove the parameters as the function documentation intends, but allowing such URLs.